Addressing feedback adding integration test.
